+++ client/src/schemas/role.schema.ts
@@ -22,40 +22,3 @@ export const GET_ROLES_WITH_USERS = gql`
     }
   }
 `;
-
-export const GET_DEPARTMENT = gql`
-  query Departments {
-    departments {
-      id
-      label
-    }
-  }
-`;
-
-export const GET_DEPARTMENT_AND_DOCTOR = gql`
-  query DepartmentsAndDoctors {
-    departments {
-      id
-      label
-    }
-    getDoctors {
-      firstname
-      id
-      lastname
-    }
-  }
-`;
-
-export const GET_DOCTORS_BY_DEPARTMENT = gql`
-  query GetDoctorByDepartment($label: String!) {
-    getDoctorByDepartment(label: $label) {
-      id
-      label
-      users {
-        firstname
-        lastname
-        id
-      }
-    }
-  }
-`;
