+++ client/src/schemas/user.schema.ts
@@ -1,0 +1,15 @@
+import { gql } from '@apollo/client';
+
+export const GET_DOCTORS_BY_DEPARTMENT = gql`
+  query GetDoctorByDepartment($label: String!) {
+    getDoctorByDepartment(label: $label) {
+      id
+      label
+      users {
+        firstname
+        lastname
+        id
+      }
+    }
+  }
+`;
