+++ core_api/src/modules/user/user.resolver.ts
@@ -1,8 +1,8 @@
-import { User, Role } from '../entities.index';
-import { Resolver, Query } from 'type-graphql';
+import { User, Role, Department } from '../entities.index';
+import { Resolver, Query, Arg } from 'type-graphql';
 
 @Resolver(User)
-export class DoctorsResolver {
+export default class UserResolver {
   @Query(() => [User], {
     description: 'Fetches all users with the role of doctor'
   })
@@ -18,5 +18,31 @@ export class DoctorsResolver {
     });
 
     return doctors;
+  }
+  @Query(() => [Department], {
+    description: 'Fetches departments by label and their doctors'
+  })
+  async getDoctorByDepartment(
+    @Arg('label', () => String) label: string
+  ): Promise<Department[]> {
+    const doctorRole = await Role.findOne({ where: { label: 'doctor' } });
+    if (!doctorRole) {
+      throw new Error("Role 'doctor' not found.");
+    }
+    const departments = await Department.find({
+      where: { label },
+      relations: ['users', 'users.role']
+    });
+
+    if (departments.length === 0) {
+      throw new Error(`No department found with the label  '${label}'.`);
+    }
+    departments.forEach((department) => {
+      department.users = department.users.filter(
+        (user) => user.role.id === doctorRole.id
+      );
+    });
+
+    return departments;
   }
 }
