+++ core_api/src/schema.ts
@@ -1,11 +1,9 @@
 import { buildSchema } from 'type-graphql';
 import RoleResolver from './modules/role/role.resolver';
-import DepartmentResolver, {
-  DoctorByDepartmentResolver
-} from './modules/department/department.resolver';
+import DepartmentResolver from './modules/department/department.resolver';
 import ConsultationResolver from './modules/consultation/consultation.resolver';
 import PatientResolver from './modules/patient/patient.resolver';
-import { DoctorsResolver } from './modules/user/user.resolver';
+import UserResolver from './modules/user/user.resolver';
 
 const getSchema = async () => {
   return await buildSchema({
@@ -14,8 +12,7 @@ const getSchema = async () => {
       ConsultationResolver,
       PatientResolver,
       DepartmentResolver,
-      DoctorByDepartmentResolver,
-      DoctorsResolver
+      UserResolver
     ],
 
     validate: true
